requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
distro==1.9.0
dj-database-url==2.3.0
Django==5.2.1
django-haystack==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.84.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyPDF2==3.0.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
whitenoise==6.9.0
Whoosh==2.7.4
